// client-app/src/app/core/enums/common-api.enum.ts
export enum CommonApiEnum {
  // user
  CurrentUser = 'users/current',
  BadgeCounters = 'users/current/badge-counters',
  Notifications = 'users/current/notifications',
  // common
  Categories = 'categories',
  Solutions = 'solutions',
  Technologies = 'technologies',
  Regions = 'regions',
  States = 'states',
  Countries = 'countries',
  Roles = 'roles',
  Permissions = 'permissions',
  HeardVia = 'heardvia',
  Timezones = 'timezones',
  RegionScaleTypes = 'get-region-scale-types',
  // articles
  Articles = 'articles',
  ArticlesSync = 'articles/sync',
  Trendings = 'trendings',
  // companies
  Companies = 'companies',
  // user pending
  UserPending = 'userpendings',
  // community
  Community = 'community',
  // tools
  Tools = 'tools',
  // users
  Users = 'users',
  // industries
  Industries = 'industries',
  // Offsite PPAs
  OffsitePPAs = 'offsiteppas',
  ContactUs = 'contact-us',
  // logging
  Logs = 'logs',
  ScheduleDemo = 'public/schedule-demo',
  GetPublicInitiativeArticle= 'get-public-initiatiative-article-content'
}